fixed errors

+++ Assignment-3/Animation.cpp
@@ -10,7 +10,7 @@ Animation::Animation(const std::string& name, const sf::Texture& t)
 {
 	m_size = Vec2(static_cast<float>(t.getSize().x) / m_frameCount, static_cast<float>(t.getSize().y));
 	m_sprite.setOrigin(m_size.x / 2.0f, m_size.y / 2.0f);
-	m_sprite.setTextureRect(sf::IntRect(m_currentFrame * m_size.x, 0, m_size.x, m_size.y));
+	m_sprite.setTextureRect(sf::IntRect(static_cast<int>(m_currentFrame * m_size.x), 0, static_cast<int>(m_size.x), static_cast<int>(m_size.y)));
 }
 
 Animation::Animation(const std::string& name, const sf::Texture& t, size_t frameCount, size_t speed)
@@ -18,14 +18,14 @@ Animation::Animation(const std::string& name, const sf::Texture& t, size_t frame
 {
 	m_size = Vec2(static_cast<float>(t.getSize().x) / m_frameCount, static_cast<float>(t.getSize().y));
 	m_sprite.setOrigin(m_size.x / 2.0f, m_size.y / 2.0f);
-	m_sprite.setTextureRect(sf::IntRect(m_currentFrame * m_size.x, 0, m_size.x, m_size.y));
+	m_sprite.setTextureRect(sf::IntRect(static_cast<int>(m_currentFrame * m_size.x), 0, static_cast<int>(m_size.x), static_cast<int>(m_size.y)));
 }
 
 void Animation::update()
 {
 	m_currentFrame++;
 	size_t frame = (m_currentFrame / m_speed) % m_frameCount;
-	m_sprite.setTextureRect(sf::IntRect(m_currentFrame * m_size.x, 0, m_size.x, m_size.y));
+	m_sprite.setTextureRect(sf::IntRect(static_cast<int>(m_currentFrame * m_size.x), 0, static_cast<int>(m_size.x), static_cast<int>(m_size.y)));
 }
 
 bool Animation::hasEnded() const
